requirements.txt
asgiref==3.6.0
backoff==2.2.1
black==23.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Deprecated==1.2.13
Django==4.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
googleapis-common-protos==1.56.2
grpcio==1.54.0
idna==3.4
importlib-metadata==6.0.1
mypy-extensions==1.0.0
opentelemetry-api==1.17.0
opentelemetry-distro==0.38b0
opentelemetry-exporter-jaeger==1.17.0
opentelemetry-exporter-jaeger-proto-grpc==1.17.0
opentelemetry-exporter-jaeger-thrift==1.17.0
opentelemetry-exporter-otlp==1.17.0
opentelemetry-exporter-otlp-proto-grpc==1.17.0
opentelemetry-exporter-otlp-proto-http==1.17.0
opentelemetry-instrumentation==0.38b0
opentelemetry-instrumentation-asgi==0.38b0
opentelemetry-instrumentation-aws-lambda==0.38b0
opentelemetry-instrumentation-dbapi==0.38b0
opentelemetry-instrumentation-django==0.38b0
opentelemetry-instrumentation-grpc==0.38b0
opentelemetry-instrumentation-logging==0.38b0
opentelemetry-instrumentation-requests==0.38b0
opentelemetry-instrumentation-sqlite3==0.38b0
opentelemetry-instrumentation-urllib==0.38b0
opentelemetry-instrumentation-urllib3==0.38b0
opentelemetry-instrumentation-wsgi==0.38b0
opentelemetry-propagator-aws-xray==1.0.1
opentelemetry-proto==1.17.0
opentelemetry-sdk==1.17.0
opentelemetry-semantic-conventions==0.38b0
opentelemetry-util-http==0.38b0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.0
protobuf==3.20.3
pytz==2023.3
requests==2.29.0
six==1.16.0
sqlparse==0.4.4
thrift==0.16.0
typing_extensions==4.5.0
urllib3==1.26.15
wrapt==1.15.0
zipp==3.15.0
